# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Finni CO2 Calculator API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects/{projectId}/versions/{versionId}/building:
    get:
      operationId: getBuilding
      tags: [building]
      summary: Building snapshot for a version
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
        - name: versionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Building data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Building"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: upsertBuilding
      tags: [building]
      summary: Upsert building (draft version only; tenant editors)
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: string
        - name: versionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertBuildingRequest"
      responses:
        "200":
          description: Building upserted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Building"
        "400":
          description: Validation or locked version
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /versions/{versionId}/products/import/preview:
    post:
      operationId: previewProductImport
      tags: [products]
      summary: Parse and validate an Excel file (draft versions only)
      parameters:
        - name: versionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required: [file]
      responses:
        "200":
          description: Parsed rows and validation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductImportPreviewResponse"
        "400":
          description: Missing file or parse error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /versions/{versionId}/products/import/commit:
    post:
      operationId: commitProductImport
      tags: [products]
      summary: Import validated rows (re-validated server-side; draft only)
      parameters:
        - name: versionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductImportCommitRequest"
      responses:
        "201":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductImportCommitResponse"
        "400":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /versions/{versionId}/bim-import/preview:
    post:
      operationId: previewBimImport
      tags: [bim]
      summary: Parse IFC STEP file (draft versions only); preview building + product mapping
      parameters:
        - name: versionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required: [file]
      responses:
        "200":
          description: Parsed IFC preview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BimImportPreviewResponse"
        "400":
          description: Missing file or parse error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /versions/{versionId}/bim-import/commit:
    post:
      operationId: commitBimImport
      tags: [bim]
      summary: Apply building and/or products from confirmed preview (re-validated server-side; draft only)
      parameters:
        - name: versionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BimImportCommitRequest"
      responses:
        "201":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BimImportCommitResponse"
        "400":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /emission-factors:
    get:
      operationId: getEmissionFactors
      tags: [emission-factors]
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: unit
          in: query
          schema:
            type: string
        - name: sourceType
          in: query
          schema:
            type: string
        - name: q
          in: query
          schema:
            type: string
          description: Case-insensitive substring match on source name and category
      responses:
        "200":
          description: List of emission factors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EmissionFactor"
...
  /tenant/emission-factors/managed:
    get:
      operationId: getTenantEmissionFactorsManaged
      tags: [tenant-epd]
      summary: List tenant-authored EPD records (admin/editor); includes archived
      responses:
        "200":
          description: Emission factor rows for this tenant only
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EmissionFactor"
        "403":
          description: Not allowed to manage tenant EPDs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /tenant/emission-factors:
    post:
      operationId: createTenantEmissionFactor
      tags: [tenant-epd]
      summary: Create tenant-specific EPD factor (sourceType EPD)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TenantEpdCreateRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmissionFactor"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /tenant/emission-factors/{id}:
    patch:
      operationId: updateTenantEmissionFactor
      tags: [tenant-epd]
      summary: Update tenant EPD record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TenantEpdUpdateRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmissionFactor"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found or not a manageable tenant EPD row
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /tenant/emission-factors/{id}/archive:
    post:
      operationId: archiveTenantEmissionFactor
      tags: [tenant-epd]
      summary: Archive (deactivate) tenant EPD record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Archived
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmissionFactor"
        "400":
          description: Already archived
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
          description: Machine-readable reason when applicable (e.g. VERSION_LOCKED, LOCK_PRECONDITIONS_FAILED).
        summary:
          type: string
        failedChecks:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              message:
                type: string
            required: [id, message]
        fieldErrors:
          type: object
          additionalProperties:
            type: string
          description: Per-field validation messages (e.g. product PATCH/POST).
      required: [error]
...
    Space:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        areaM2:
          type: number
      required: [id, name, areaM2]
    Building:
      type: object
      properties:
        id:
          type: string
        versionId:
          type: string
        grossAreaM2:
          type: number
          nullable: true
        spaces:
          type: array
          items:
            $ref: "#/components/schemas/Space"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, versionId, spaces, createdAt, updatedAt]
    UpsertBuildingRequest:
      type: object
      properties:
        grossAreaM2:
          type: number
          description: Required on save; must be non-negative
        spaces:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              name:
                type: string
              areaM2:
                type: number
            required: [name, areaM2]
      required: [grossAreaM2, spaces]
...
    ProductImportRow:
      type: object
      description: Validated product row (same rules as create + module shares)
      properties:
        name:
          type: string
        category:
          type: string
        quantityValue:
          type: number
          nullable: true
        quantityUnit:
          type: string
          nullable: true
        moduleA1A3Share:
          type: number
        moduleA4Share:
          type: number
        moduleA5Share:
          type: number
        moduleBShare:
          type: number
        moduleCShare:
          type: number
      required:
        [name, category, moduleA1A3Share, moduleA4Share, moduleA5Share, moduleBShare, moduleCShare]
    ProductImportPreviewRow:
      type: object
      properties:
        excelRow:
          type: integer
        ok:
          type: boolean
        fieldErrors:
          type: object
          additionalProperties:
            type: string
        data:
          $ref: "#/components/schemas/ProductImportRow"
      required: [excelRow, ok]
    ProductImportPreviewResponse:
      type: object
      properties:
        worksheetName:
          type: string
        structureOk:
          type: boolean
        structureError:
          type: string
          nullable: true
        code:
          type: string
          nullable: true
        rows:
          type: array
          items:
            $ref: "#/components/schemas/ProductImportPreviewRow"
        validCount:
          type: integer
        errorCount:
          type: integer
      required: [worksheetName, structureOk, rows, validCount, errorCount]
    ProductImportCommitRequest:
      type: object
      properties:
        rows:
          type: array
          items:
            $ref: "#/components/schemas/ProductImportRow"
      required: [rows]
    ProductImportCommitResponse:
      type: object
      properties:
        imported:
          type: integer
        productIds:
          type: array
          items:
            type: string
      required: [imported, productIds]
    BimImportPreviewSpace:
      type: object
      properties:
        sourceExpressId:
          type: integer
        name:
          type: string
        areaM2:
          type: number
          nullable: true
      required: [sourceExpressId, name]
    BimImportPreviewProductRow:
      type: object
      properties:
        sourceExpressId:
          type: integer
        ifcType:
          type: string
        ok:
          type: boolean
        fieldErrors:
          type: object
          additionalProperties:
            type: string
        data:
          $ref: "#/components/schemas/ProductImportRow"
      required: [sourceExpressId, ifcType, ok]
    BimImportPreviewResponse:
      type: object
      properties:
        structureOk:
          type: boolean
        structureError:
          type: string
          nullable: true
        schemaName:
          type: string
          nullable: true
        scopeNotes:
          type: array
          items:
            type: string
        buildingName:
          type: string
          nullable: true
        suggestedGrossAreaM2:
          type: number
          nullable: true
        spaces:
          type: array
          items:
            $ref: "#/components/schemas/BimImportPreviewSpace"
        warnings:
          type: array
          items:
            type: string
        productRows:
          type: array
          items:
            $ref: "#/components/schemas/BimImportPreviewProductRow"
        stats:
          type: object
          properties:
            spaceCount:
              type: integer
            productCandidates:
              type: integer
            productOk:
              type: integer
            productError:
              type: integer
          required: [spaceCount, productCandidates, productOk, productError]
      required: [structureOk, scopeNotes, spaces, warnings, productRows, stats]
    BimImportCommitBuilding:
      type: object
      properties:
        grossAreaM2:
          type: number
        spaces:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              areaM2:
                type: number
            required: [name, areaM2]
      required: [grossAreaM2, spaces]
    BimImportCommitRequest:
      type: object
      properties:
        applyBuilding:
          type: boolean
        building:
          $ref: "#/components/schemas/BimImportCommitBuilding"
        products:
          type: array
          items:
            $ref: "#/components/schemas/ProductImportRow"
      required: [applyBuilding, products]
    BimImportCommitResponse:
      type: object
      properties:
        importedProducts:
          type: integer
        productIds:
          type: array
          items:
            type: string
        buildingUpdated:
          type: boolean
      required: [importedProducts, productIds, buildingUpdated]
...
    EmissionFactor:
      type: object
      properties:
        id:
          type: string
        tenantId:
          type: string
          nullable: true
          description: Null = platform-wide catalog; set = tenant-owned factor
        sourceType:
          type: string
          enum: [generic, EPD, product, external]
        sourceName:
          type: string
        category:
          type: string
        unit:
          type: string
          enum: [kg, m2, m3]
        co2ePerUnit:
          type: number
        active:
          type: boolean
        externalSourceKey:
          type: string
          nullable: true
        externalRecordId:
          type: string
          nullable: true
        lastSyncedAt:
          type: string
          format: date-time
          nullable: true
      required: [id, sourceType, sourceName, category, unit, co2ePerUnit, active]
...
    TenantEpdCreateRequest:
      type: object
      properties:
        sourceName:
          type: string
        category:
          type: string
        unit:
          type: string
          enum: [kg, m2, m3]
        co2ePerUnit:
          type: number
          minimum: 0
      required: [sourceName, category, unit, co2ePerUnit]
    TenantEpdUpdateRequest:
      type: object
      properties:
        sourceName:
          type: string
        category:
          type: string
        unit:
          type: string
          enum: [kg, m2, m3]
        co2ePerUnit:
          type: number
          minimum: 0
...